requirements.txt
altgraph==0.17.3
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2021.5.30
chardet==4.0.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
customtkinter==5.1.2
cvxopt==1.3.2
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.3.3
fonttools==4.41.1
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
MouseInfo==0.1.3
nest-asyncio==1.5.6
numpy==1.24.3
openpyxl==3.1.2
packaging==23.0
pandas==2.0.2
parso==0.8.3
pathlib==1.0.1
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
PyAutoGUI==0.9.53
pyFirmata==1.1.0
PyGetWindow==0.0.9
Pygments==2.14.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2023.2
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==25.0.2
requests==2.25.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
threadpoolctl==3.2.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.10.0
tzdata==2023.3
urllib3==1.26.6
wcwidth==0.2.6
Werkzeug==2.3.7
zipp==3.15.0
